Fix mypy error and import for test
