+++ providers/amazon/tests/unit/amazon/aws/cli/test_definition.py
